refactor: Rename forgotten file
