+++ core/credit/src/ledger/templates/mod.rs
@@ -7,8 +7,8 @@ mod create_credit_facility;
 mod initiate_disbursal;
 mod obligation_defaulted_balance;
 mod obligation_due_balance;
+mod obligation_fulfillment;
 mod obligation_overdue_balance;
-mod payment_allocation;
 mod post_accrued_interest;
 mod remove_collateral;
 mod reserve_for_liquidation;
@@ -22,8 +22,8 @@ pub use create_credit_facility::*;
 pub use initiate_disbursal::*;
 pub use obligation_defaulted_balance::*;
 pub use obligation_due_balance::*;
+pub use obligation_fulfillment::*;
 pub use obligation_overdue_balance::*;
-pub use payment_allocation::*;
 pub use post_accrued_interest::*;
 pub use remove_collateral::*;
 pub use reserve_for_liquidation::*;
