+++ core/credit/src/ledger/templates/obligation_fulfillment.rs
